Move new tests to prognostic run diags
