+++ workflows/prognostic_run_diags/fv3net/diagnostics/prognostic_run/transform.py
@@ -173,7 +173,6 @@ def daily_mean(split: timedelta, arg: DiagArg) -> DiagArg:
         arg: input arguments to transform prior to the diagnostic calculation
     """
     prognostic, verification, grid = arg.prediction, arg.verification, arg.grid
-
     split_time = prognostic.time.values[0] + split
     prognostic = _resample_end(prognostic, split_time, "1D")
     verification = _resample_end(verification, split_time, "1D")
